remove unused WETH and fix remaining lint issues

+++ src/ReferralDepositAdapter.sol
@@ -85,7 +85,7 @@ contract ReferralDepositAdapter is
     /// @notice Publishes multiple referral information using the existing event.
     /// @param referrals Array of ReferralInfo structs containing referral details.
     function publishReferrals(ReferralInfo[] calldata referrals) external  onlyRole(REFERRAL_PUBLISHER_ROLE)  {
-        for (uint i = 0; i < referrals.length; i++) {
+        for (uint256 i = 0; i < referrals.length; i++) {
             emit ReferralDepositProcessed(
                 referrals[i].depositor,
                 referrals[i].referee,
